Linux and OS-less working mode, new message buffer and FIFO
- new buffer structure that supports message aging to prevent clogging
- the TX timestamp write back and processing have been moved into task_ptp.c
- ptp_transmit_enqueue() may be called inside an interrupt
- the heartbeat timer has been moved into the task_ptp.c moving all OS-aware functionality into the task_ptp.c file
- new Linux NSD; bundled NSDs have been updated
- CMakeLists.txt have been updated
- new logging feature: TRANSMISSION logging
- "%lu" formatting markers replaced with the __PRI64_PREFIX macro
- task_ptp() methods and functions have been extended with conditional compiling on Linux and OS-less platforms
- skipping synchronization cycles at 0 time error fixed
- the documentation has been updated and examples have been added

+++ src/flexptp/profiles.c
@@ -19,7 +19,7 @@ void ptp_print_profile() {
         PTP_TRANSPEC_HINT[ptp_get_transport_specific()],
         PTP_TRANSPORT_TYPE_HINT[ptp_get_transport_type()],
         PTP_DELMECH_HINT[ptp_get_delay_mechanism()],
-        ptp_get_domain());
+        (uint32_t)ptp_get_domain());
 
     int8_t drlp = ptp_get_delay_req_log_period();
     if (drlp != PTP_LOGPER_SYNCMATCHED) {
@@ -39,11 +39,11 @@ void ptp_print_profile() {
 
     uint8_t flags = ptp_get_profile_flags();
     MSG(PTP_COLOR_BYELLOW " flags:              " PTP_COLOR_CYAN "0x%X\n", flags);
-    for (uint8_t i = 0; i < PTP_PF_N - 1; i++) {
+    for (uint32_t i = 0; i < PTP_PF_N - 1; i++) {
         uint8_t mask = 1 << i;
         bool on = (flags & mask) != 0;
         const char * color = on ? (PTP_COLOR_BGREEN) : (PTP_COLOR_BRED);
         char bullet = on ? '+' : '-';
-        MSG("   %s%c (0x%X) %s" PTP_COLOR_RESET "\n", color, bullet, mask,  PTP_FLAGS_HINT[i]);
+        MSG("   %s%c (0x%X) %s" PTP_COLOR_RESET "\n", color, bullet, (uint32_t)mask, PTP_FLAGS_HINT[i]);
     }
 }
